Move passport logic out of index
